+++ swifttools/swift_too/common.py
@@ -62,7 +62,7 @@ modesxrt = {
     "Bias": 9,
     "PC_150": 150,
     "PC_200": 200,
-    "Manual": 255
+    "Manual": 255,
 }
 
 # Regex for matching date, time and datetime strings
@@ -90,7 +90,12 @@ def convert_to_dt(value, isutc=False, outfunc=datetime):
         dtvalue = outfunc.strptime(f"{value} 00:00:00", "%Y-%m-%d %H:%M:%S")
     elif type(value) == outfunc:
         dtvalue = value  # Just pass through un molested
-    elif type(value) == swiftdatetime and outfunc == datetime or type(value) == datetime and outfunc == swiftdatetime:
+    elif (
+        type(value) == swiftdatetime
+        and outfunc == datetime
+        or type(value) == datetime
+        and outfunc == swiftdatetime
+    ):
         dtvalue = outfunc.fromtimestamp(value.timestamp())
     elif value is None:
         dtvalue = None
@@ -532,7 +537,11 @@ class TOOAPI_Daterange:
 
     @property
     def begin(self):
-        if hasattr(self._begin, "utcf") and self._begin.utcf is None and self.utcf is not None:
+        if (
+            hasattr(self._begin, "utcf")
+            and self._begin.utcf is None
+            and self.utcf is not None
+        ):
             self._begin.utcf = self.utcf
         return self._begin
 
@@ -540,7 +549,11 @@ class TOOAPI_Daterange:
     def end(self):
         if self._begin is not None and self._length is not None and self._end is None:
             self._end = self.begin + timedelta(days=self._length)
-        if hasattr(self._end, "utcf") and self._end.utcf is None and self.utcf is not None:
+        if (
+            hasattr(self._end, "utcf")
+            and self._end.utcf is None
+            and self.utcf is not None
+        ):
             self._end.utcf = self.utcf
         return self._end